Fix order dependent component registration issue
Due to the way Go and this library works, it's necessary to know what
components exist vs what is just HTML in the templat that looks like
a template.

This makes some trade-offs, specifically requiring that we spend
redundant time re-generating the template when a component is
registered after the template is parsed.

Additionally, we hold onto the raw content of the template so we can
recompile it if necessary.