Introduce dynamic func maps per-render
Applications generally need to provide request specific data to the UI,
which is often nested deep in template rendering hierarchies. This
change introduces a new method, `RenderWithFuncs`, which accepts a
`glam.FuncMap` that can override methods in the default `FuncMap` passed
to the engine.

This should allow applications to pass any request specific data to the
templates via functions, like CSRFs, current user, etc.

This might be less ideal than using generics to define a typed request
object that could be passed to the templates, but this is simpler for
the time being.